navit: Remove Navit & everything that depends on it
Most of these components are unmaintained and it already has created
issues for people:

https://at.projects.genivi.org/jira/browse/GDP-324

And now that we want to upgrade our stack, it'll be difficult to keep
unmaintained software against new components and newer (more strick)
compilers.

This patch also drops "Mock Navigation" app but that doesn't seem to be
anything that impresses anyone anyway so I wouldn't shed any tears on it's
loss.